third_party/blink/web_tests/platform/generic/fast/events/tab-focus-link-in-canvas-expected.txt

Ensures that you can tab to focus a link inside a canvas subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement.id is "before"
PASS document.activeElement.id is "canvas_link"
PASS successfullyParsed is true

TEST COMPLETE

Before 
